fix: Stop deadlock when any NsAction fails
When an NsAction fails, it pushes the error to the routine's error
channel. However, before this would not return after, therefore when
the final nil value would try to be pushed to the chan, this would
casue a deadlock (since it is a buffered channel). To fix this, a
return is called after any NsAction fail, ensuring that only one err
can ever be pushed to the goroutine.